factor out factory builder functions
