+++ src/Model/Componentfactory/Componentfactory.cpp
@@ -1,18 +1,5 @@
 #include <Componentfactory.hpp>
 
-#include "Compressorstation.hpp"
-#include "Controlvalve.hpp"
-#include "Gaspowerconnection.hpp"
-#include "Innode.hpp"
-#include "PQnode.hpp"
-#include "PVnode.hpp"
-#include "Pipe.hpp"
-#include "Shortpipe.hpp"
-#include "Sink.hpp"
-#include "Source.hpp"
-#include "Transmissionline.hpp"
-#include "Vphinode.hpp"
-
 namespace Model::Componentfactory {
 
   // std::map<std::string, std::unique_ptr<AbstractNodeType>> Componentfactory::get_node_type_map() {
@@ -55,29 +42,7 @@ namespace Model::Componentfactory {
     return topology_schema;
   }
 
-  void add_gas_components(Componentfactory &factory) {
-    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Gas::Source>>());
-    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Gas::Sink>>());
-    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Gas::Innode>>());
 
-    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Pipe>>());
-    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Shortpipe>>());
-    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Controlvalve>>());
-    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Compressorstation>>());
-    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Gaspowerconnection>>());
-  }
 
-  void add_power_components(Componentfactory &factory) {
-    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Power::Vphinode>>());
-    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Power::PVnode>>());
-    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Power::PQnode>>());
-
-    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Power::Transmissionline>>());
-  }
-
-  void add_all_components(Componentfactory &factory){
-    add_power_components(factory);
-    add_gas_components(factory);
-  }
 
 }//namespace
