+++ src/Model/Componentfactory/Gas_factory/Gas_factory.cpp
@@ -11,6 +11,18 @@
 
 namespace Model::Componentfactory {
 
+  void add_gas_components(Componentfactory &factory) {
+    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Gas::Source>>());
+    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Gas::Sink>>());
+    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Gas::Innode>>());
+
+    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Pipe>>());
+    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Shortpipe>>());
+    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Controlvalve>>());
+    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Compressorstation>>());
+    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Gas::Gaspowerconnection>>());
+  }
+
   std::map<std::string, Nodefactory> Gas_factory::get_nodetypemap_impl(){
 
     std::map<std::string, Nodefactory> data;
