+++ src/Model/Componentfactory/Gas_factory/include/Gas_factory.hpp
@@ -4,6 +4,12 @@
 
 namespace Model::Componentfactory {
 
+  /**
+   * @brief add Gas Components to the provided Componentfactory
+   * @param factory to which the Gas Component schould be added
+   */
+  void add_gas_components(Componentfactory &factory);
+
   /// \brief Provides factories for gas component types.
   struct Gas_factory : public Componentfactory_interface<Gas_factory> {
