+++ src/Model/Componentfactory/Power_factory/Power_factory.cpp
@@ -7,6 +7,14 @@
 
 namespace Model::Componentfactory {
 
+  void add_power_components(Componentfactory &factory) {
+    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Power::Vphinode>>());
+    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Power::PVnode>>());
+    factory.add_node_type(std::make_unique<NodeType<Networkproblem::Power::PQnode>>());
+
+    factory.add_edge_type(std::make_unique<EdgeType<Networkproblem::Power::Transmissionline>>());
+  }
+
   PowerFactory::PowerFactory() {
     std::vector<std::unique_ptr<AbstractNodeType>> node_types;
