+++ src/Model/Componentfactory/Power_factory/include/Power_factory.hpp
@@ -4,6 +4,13 @@
 
 namespace Model::Componentfactory {
 
+  /**
+   * @brief add the Power Components to the provided Componentfactory
+   * 
+   * @param factory to which to add the Power Components
+   */
+  void add_power_components(Componentfactory &factory);
+
   struct PowerFactory : public Componentfactory {
     PowerFactory();
   };
